requirements.txt
langchain
pypdf
openai
faiss-cpu
python-dotenv
tiktoken
streamlit